+++ crates/wasmparser/src/limits.rs
@@ -65,7 +65,7 @@ pub fn max_wasm_memory64_pages(page_size: u64) -> u64 {
 pub use self::component_limits::*;
 #[cfg(feature = "component-model")]
 mod component_limits {
-    pub const MAX_WASM_MODULE_SIZE: usize = 1024 * 1024 * 1024; //= 1 GiB
+    pub const MAX_WASM_MODULE_SIZE: u32 = 1024 * 1024 * 1024; //= 1 GiB
     pub const MAX_WASM_MODULE_TYPE_DECLS: usize = 100_000;
     pub const MAX_WASM_COMPONENT_TYPE_DECLS: usize = 1_000_000;
     pub const MAX_WASM_INSTANCE_TYPE_DECLS: usize = 1_000_000;
