+++ crates/wasmparser/src/readers.rs
@@ -113,13 +113,13 @@ impl<'a, T> SectionLimited<'a, T> {
     }
 
     /// Returns whether the original byte offset of this section.
-    pub fn original_position(&self) -> usize {
+    pub fn original_position(&self) -> u64 {
         self.reader.original_position()
     }
 
     /// Returns the range, as byte offsets, of this section within the original
     /// wasm binary.
-    pub fn range(&self) -> Range<usize> {
+    pub fn range(&self) -> Range<u64> {
         self.reader.range()
     }
 
@@ -182,7 +182,7 @@ pub struct SectionLimitedIntoIter<'a, T> {
 
 impl<T> SectionLimitedIntoIter<'_, T> {
     /// Returns the current byte offset of the section within this iterator.
-    pub fn original_position(&self) -> usize {
+    pub fn original_position(&self) -> u64 {
         self.section.reader.original_position()
     }
 }
@@ -230,7 +230,7 @@ impl<'a, T> Iterator for SectionLimitedIntoIterWithOffsets<'a, T>
 where
     T: FromReader<'a>,
 {
-    type Item = Result<(usize, T)>;
+    type Item = Result<(u64, T)>;
 
     fn next(&mut self) -> Option<Self::Item> {
         let pos = self.iter.section.reader.original_position();
@@ -277,13 +277,13 @@ impl<'a, T> Subsections<'a, T> {
     }
 
     /// Returns whether the original byte offset of this section.
-    pub fn original_position(&self) -> usize {
+    pub fn original_position(&self) -> u64 {
         self.reader.original_position()
     }
 
     /// Returns the range, as byte offsets, of this section within the original
     /// wasm binary.
-    pub fn range(&self) -> Range<usize> {
+    pub fn range(&self) -> Range<u64> {
         self.reader.range()
     }
